// mortware/src/app/Models/company.model.ts
export interface ICompanyModel {

    companyId: number; 
 nameOfficial: string;
 nameShort: string;
 currency1: string;
currency2: string;
 rateCur1toCur2?:number;
 phone: string;
 fax: string;
 email1: string;
 email2: string;
 mail?: boolean;
 nodeKey: string;
 openImage?: number;
 closeImage?: number;
 website: string;
 controllingPerson: string;
 companyAddressId?: number;
 phone2: string;
 employerIdentificationNumber: string;
}


 